@@ -4,14 +4,14 @@ version: 2.0.0
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0-417.1.beta <4.0.0"
+  sdk: ^3.1.0
+  flutter: ">=3.1.0 <4.0.0"
 
 dependencies:
-  cache:
-    path: ../cache
   equatable: ^2.0.5
   firebase_auth: ^4.7.1
   firebase_core: ^2.15.0
+  cloud_firestore: ^4.9.1
   flutter:
     sdk: flutter
   google_sign_in: ^6.1.4
