@@ -3,7 +3,8 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0-417.1.beta <4.0.0"
+  sdk: ^3.1.0
+  flutter: ">=3.1.0 <4.0.0"
 
 dependencies:
   formz: ^0.6.0
