@@ -1,10 +1,11 @@
 name: budget_app
 description: A Budget app.
-version: 2.6.0
+version: 2.5.0
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0-417.1.beta <4.0.0"
+  sdk: ^3.1.0
+  flutter: ">=3.1.0 <4.0.0"
 
 dependencies:
   intl:
@@ -18,7 +19,7 @@ dependencies:
   google_fonts: ^5.1.0
   meta: ^1.7.0
   uuid: ^3.0.0
-  json_annotation: ^4.8.0
+  json_annotation: ^4.8.1
   http: ^1.1.0
   chart:
     path: packages/chart
@@ -26,16 +27,23 @@ dependencies:
     path: packages/authentication_repository
   form_inputs:
     path: packages/form_inputs
+  network:
+    path: packages/network
+  cache_client:
+    path: packages/cache_client
   bloc_concurrency: ^0.2.1
-  shared_preferences: ^2.1.2
   firebase_core: ^2.15.0
   firebase_auth: ^4.7.1
+  cloud_firestore: ^4.9.3
   carousel_slider: ^4.2.1
   adaptive_breakpoints: ^0.1.6
-  fl_chart: ^0.63.0
   font_awesome_flutter: ^10.5.0
   flow_builder: ^0.0.10
   flutter_native_splash: ^2.3.7
+  go_router: ^13.2.0
+  hydrated_bloc: ^9.1.5
+  path_provider: ^2.1.2
+  animations: ^2.0.11
 
 dev_dependencies:
   bloc_test: ^9.0.0
