@@ -2,63 +2,48 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:bloc_concurrency/bloc_concurrency.dart';
-import 'package:budget_app/constants/api.dart';
-import 'package:budget_app/transactions/models/transaction_tile.dart';
-import 'package:budget_app/transactions/repository/transactions_repository.dart';
-import 'package:budget_app/transfer/transfer.dart';
+import 'package:budget_app/budgets/budgets.dart';
+import 'package:cache_client/cache_client.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
 import 'package:form_inputs/form_inputs.dart';
 import 'package:formz/formz.dart';
 
-import '../../accounts/models/account.dart';
-import '../../accounts/repository/accounts_repository.dart';
-import '../../categories/models/category.dart';
-import '../../categories/repository/categories_repository.dart';
+import '../../accounts_list/models/account.dart';
+import '../../budgets/repository/budget_repository.dart';
 import '../../constants/constants.dart';
-import '../../transactions/models/transaction_type.dart';
+import '../../transaction/transaction.dart';
 
 part 'transfer_event.dart';
-
 part 'transfer_state.dart';
 
 class TransferBloc extends Bloc<TransferEvent, TransferState> {
   final TransactionsRepository _transactionsRepository;
-  final CategoriesRepository _categoriesRepository;
-  late final AccountsRepository _accountsRepository;
-  late final StreamSubscription<List<Account>> _accountsSubscription;
-  late final StreamSubscription<List<Category>> _categoriesSubscription;
+  final BudgetRepository _budgetRepository;
+  late final StreamSubscription<Budget> _budgetSubscription;
 
   TransferBloc(
       {required TransactionsRepository transactionsRepository,
-      required CategoriesRepository categoriesRepository,
-      required AccountsRepository accountsRepository})
+      required BudgetRepository budgetRepository})
       : _transactionsRepository = transactionsRepository,
-        _categoriesRepository = categoriesRepository,
-        _accountsRepository = accountsRepository,
+        _budgetRepository = budgetRepository,
         super(TransferState()) {
+    _budgetSubscription = _budgetRepository.budget.listen((budget) {
+      add(TransferBudgetChanged(budget: budget));
+    });
     on<TransferEvent>(_onEvent, transformer: sequential());
-    _accountsSubscription =
-        _accountsRepository.getAccounts().skip(1).listen((accounts) {
-      add(TransferAccountsChanged(accounts: accounts));
-    });
-    _categoriesSubscription =
-        _categoriesRepository.getCategories().skip(1).listen((categories) {
-      add(TransferCategoriesChanged(categories: categories));
-    });
   }
 
   Future<void> _onEvent(
       TransferEvent event, Emitter<TransferState> emit) async {
     return switch (event) {
+      final TransferBudgetChanged e => _onBudgetChanged(e, emit),
       final TransferFormLoaded e => _onFormLoaded(e, emit),
       final TransferAmountChanged e => _onAmountChanged(e, emit),
       final TransferDateChanged e => _onDateChanged(e, emit),
-      final TransferAccountsChanged e => _onAccountsChanged(e, emit),
       final TransferFromAccountChanged e => _onFromAccountChanged(e, emit),
       final TransferToAccountChanged e => _onToAccountChanged(e, emit),
       final TransferNotesChanged e => _onNotesChanged(e, emit),
-      final TransferCategoriesChanged e => _onCategoriesChanged(e, emit),
       final TransferFormSubmitted e => _onFormSubmitted(e, emit),
     };
   }
@@ -66,45 +51,34 @@ class TransferBloc extends Bloc<TransferEvent, TransferState> {
   Future<void> _onFormLoaded(
       TransferFormLoaded event, Emitter<TransferState> emit) async {
     emit(state.copyWith(trStatus: TransferStatus.loading));
-    final accounts = await _accountsRepository.getAccounts().first;
-    final accCategories = await _categoriesRepository.getCategories().first;
-    final filteredCategories = accCategories
-        .where((aC) => aC.transactionType == TransactionType.ACCOUNT)
-        .toList();
-    final transactionTile = event.transactionTile;
-    var id;
-    var fromAccount;
-    var toAccount;
-    if (transactionTile != null) {
-      id = transactionTile.id;
-      fromAccount = accounts
-          .where((acc) => acc.id == transactionTile.fromAccount!.id)
-          .first;
-      toAccount = accounts
-          .where((acc) => acc.id == transactionTile.toAccount!.id)
-          .first;
+    final transaction = event.transaction;
+    String? id;
+    Account? fromAccount;
+    Account? toAccount;
+    if (transaction != null) {
+      id = transaction.id;
+      fromAccount = transaction.fromAccount!;
+      toAccount = transaction.toAccount!;
     }
     await Future.delayed(Duration(milliseconds: 100));
     emit(state.copyWith(
-      budgetId: await getBudgetId(),
+      editedTransfer: transaction,
       id: id,
-      amount: transactionTile == null
+      amount: transaction == null
           ? Amount.pure()
-          : Amount.dirty(transactionTile.amount.toString()),
+          : Amount.dirty(transaction.amount.toString()),
       trStatus: TransferStatus.success,
       fromAccount: fromAccount,
       toAccount: toAccount,
-      date: transactionTile?.dateTime ?? DateTime.now(),
-      notes: transactionTile?.description ?? '',
-      accountCategories: filteredCategories,
-      accounts: accounts,
-      isValid: transactionTile == null ? false : true,
+      date: transaction?.dateTime ?? DateTime.now(),
+      notes: transaction?.description ?? '',
+      isValid: transaction == null ? false : true,
     ));
   }
 
-  void _onCategoriesChanged(
-      TransferCategoriesChanged event, Emitter<TransferState> emit) {
-    emit(state.copyWith(accountCategories: event.categories));
+  void _onBudgetChanged(
+      TransferBudgetChanged event, Emitter<TransferState> emit) {
+    emit(state.copyWith(budget: event.budget));
   }
 
   void _onAmountChanged(
@@ -122,11 +96,6 @@ class TransferBloc extends Bloc<TransferEvent, TransferState> {
     emit(
       state.copyWith(date: event.dateTime),
     );
-  }
-
-  void _onAccountsChanged(
-      TransferAccountsChanged event, Emitter<TransferState> emit) {
-    emit(state.copyWith(accounts: event.accounts));
   }
 
   void _onFromAccountChanged(
@@ -147,21 +116,22 @@ class TransferBloc extends Bloc<TransferEvent, TransferState> {
   Future<void> _onFormSubmitted(
       TransferFormSubmitted event, Emitter<TransferState> emit) async {
     emit(state.copyWith(status: FormzSubmissionStatus.inProgress));
-    final transfer = Transfer(
+    final transfer = Transaction(
       id: state.id,
-      budgetId: state.budgetId!,
+      type: TransactionType.TRANSFER,
       amount: double.parse(state.amount.value),
       date: state.date ?? DateTime.now(),
-      fromAccountId: state.fromAccount!.id!,
-      toAccountId: state.toAccount!.id!,
-      notes: state.notes,
+      fromAccountId: state.fromAccount!.id,
+      toAccountId: state.toAccount!.id,
+      description: state.notes,
+      budgetId: await CacheClient.instance.getBudgetId() ?? '',
     );
     try {
-      if (state.id == null) {
-        await _transactionsRepository.createTransfer(transfer);
-      } else {
-        await _transactionsRepository.editTransfer(transfer);
-      }
+      await state.editedTransfer == null
+          ? _transactionsRepository.createTransaction(transfer)
+          : _transactionsRepository.updateTransaction(transfer);
+      _updateBudgetOnTransfer(
+          transfer: transfer, editedTransfer: state.editedTransfer);
       emit(state.copyWith(status: FormzSubmissionStatus.success));
       isDisplayDesktop(event.context!)
           ? add(TransferFormLoaded())
@@ -171,10 +141,54 @@ class TransferBloc extends Bloc<TransferEvent, TransferState> {
     }
   }
 
+  void _updateBudgetOnTransfer(
+      {required Transaction transfer,
+      ComprehensiveTransaction? editedTransfer}) {
+    List<Account> updatedAccounts = [];
+
+    final accounts = state.budget.accountList;
+    //find the acc from editedTransaction and return amount
+    //find the acc from transaction and update amount
+    if (editedTransfer != null) {
+      updatedAccounts = accounts.map((acc) {
+        if (acc.id == editedTransfer.fromAccount!.id) {
+          return acc.copyWith(balance: acc.balance + editedTransfer.amount);
+        } else {
+          return acc;
+        }
+      }).toList();
+      updatedAccounts = updatedAccounts.map((acc) {
+        if (acc.id == editedTransfer.toAccount!.id) {
+          return acc.copyWith(balance: acc.balance - editedTransfer.amount);
+        } else {
+          return acc;
+        }
+      }).toList();
+    } else {
+      updatedAccounts = [...accounts];
+    }
+
+    updatedAccounts = updatedAccounts.map((acc) {
+      if (acc.id == transfer.fromAccountId) {
+        return acc.copyWith(balance: acc.balance - transfer.amount);
+      } else {
+        return acc;
+      }
+    }).toList();
+    updatedAccounts = updatedAccounts.map((acc) {
+      if (acc.id == transfer.toAccountId) {
+        return acc.copyWith(balance: acc.balance + transfer.amount);
+      } else {
+        return acc;
+      }
+    }).toList();
+
+    _budgetRepository.pushUpdatedAccounts(updatedAccounts);
+  }
+
   @override
   Future<void> close() {
-    _accountsSubscription.cancel();
-    _categoriesSubscription.cancel();
+    _budgetSubscription.cancel();
     return super.close();
   }
 }
