@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../../../../utils/theme/cubit/theme_cubit.dart';
 import '../../debt_form.dart';
 
 class AprInputField extends StatelessWidget {
@@ -8,6 +9,7 @@ class AprInputField extends StatelessWidget {
   Widget build(BuildContext context) {
     return BlocBuilder<DebtBloc, DebtState>(
       builder: (context, state) {
+        final themeState = context.read<ThemeCubit>().state;
         return TextFormField(
           initialValue: state.apr.value,
           onChanged: (apr) => context
@@ -17,7 +19,7 @@ class AprInputField extends StatelessWidget {
           decoration: InputDecoration(
             icon: Icon(
               Icons.percent,
-              color: Colors.orangeAccent,
+              color: themeState.secondaryColor,
             ),
             border: OutlineInputBorder(),
             labelText: 'APR',
