@@ -1,11 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../../../../utils/theme/cubit/theme_cubit.dart';
 import '../../debt_form.dart';
 
 class BalanceInput extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    final themeState = context.read<ThemeCubit>().state;
     return BlocBuilder<DebtBloc, DebtState>(
       builder: (context, state) {
         return TextFormField(
@@ -17,7 +19,7 @@ class BalanceInput extends StatelessWidget {
           decoration: InputDecoration(
             icon: Icon(
               Icons.attach_money,
-              color: Colors.orangeAccent,
+              color: themeState.secondaryColor,
             ),
             border: OutlineInputBorder(),
             labelText: 'Balance',
