@@ -1,11 +1,13 @@
-import 'package:budget_app/account_edit/bloc/account_edit_bloc.dart';
 import 'package:budget_app/debt_payoff_planner/debt_form/bloc/debt_bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../../../../utils/theme/cubit/theme_cubit.dart';
+
 class NameInputField extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    final themeState = context.read<ThemeCubit>().state;
     return BlocBuilder<DebtBloc, DebtState>(
       buildWhen: (previous, current) => previous.name != current.name,
       builder: (context, state) {
@@ -14,7 +16,7 @@ class NameInputField extends StatelessWidget {
             decoration: InputDecoration(
               icon: Icon(
                 Icons.notes,
-                color: Colors.orangeAccent,
+                color: themeState.secondaryColor,
               ),
               border: OutlineInputBorder(),
               labelText: 'Name',
