@@ -1,10 +1,10 @@
-import 'package:budget_app/constants/colors.dart';
 import 'package:budget_app/debt_payoff_planner/cubits/debt_cubit/debts_cubit.dart';
-import 'package:budget_app/utils/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
 
+import '../../../utils/theme/budget_theme.dart';
+import '../../../utils/theme/cubit/theme_cubit.dart';
 import '../../models/debt.dart';
 
 class DebtTile extends StatelessWidget {
@@ -15,7 +15,7 @@ class DebtTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final scheme = Theme.of(context).colorScheme;
+    final themeState = context.read<ThemeCubit>().state;
     return Card(
       margin: EdgeInsets.all(10),
       child: Column(
@@ -24,45 +24,36 @@ class DebtTile extends StatelessWidget {
               padding: EdgeInsets.only(top: 0, left: 15, right: 0, bottom: 0),
               width: double.infinity,
               decoration: BoxDecoration(
-                  borderRadius: BorderRadius.only(
-                      topRight: Radius.circular(10.0),
-                      bottomRight: Radius.zero,
-                      topLeft: Radius.circular(10.0),
-                      bottomLeft: Radius.zero),
-                  color: BudgetTheme.isDarkMode(context)
-                      ? BudgetColors.primary
-                      : BudgetColors.primary600),
+                borderRadius: BorderRadius.only(
+                    topRight: Radius.circular(10.0),
+                    bottomRight: Radius.zero,
+                    topLeft: Radius.circular(10.0),
+                    bottomLeft: Radius.zero),
+                color: themeState.primaryColor[600]
+              ),
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
                   Column(
                     children: [
                       Text(debtModel.name,
-                          style: Theme.of(context)
-                              .textTheme
-                              .titleMedium!
-                              .copyWith(color: Colors.white)),
+                          style: Theme.of(context).textTheme.titleMedium!.copyWith(color: Colors.white)),
                       Text(
                         '${debtModel.apr} % APR',
-                        style: Theme.of(context)
-                            .textTheme
-                            .bodySmall!
-                            .copyWith(color: Colors.white),
+                        style: Theme.of(context).textTheme.bodySmall!.copyWith(color: Colors.white),
                       ),
                     ],
                   ),
                   ButtonBar(
                     children: [
                       IconButton.outlined(
-                          color: BudgetColors.accent,
+                        color: Colors.white,
                           onPressed: () => onEdit(debtModel),
                           icon: const Icon(Icons.edit_note)),
                       IconButton.outlined(
-                          color: BudgetColors.accent,
+                          color: Colors.white,
                           onPressed: () {
-                            context
-                                .read<DebtsCubit>()
-                                .deleteDebt(debtModel.id!);
+                            context.read<DebtsCubit>().deleteDebt(debtModel.id!);
                             //context.read<StrategyCubit>().fetchStrategy();
                           },
                           icon: const Icon(Icons.delete)),
@@ -73,8 +64,8 @@ class DebtTile extends StatelessWidget {
           Expanded(
             child: Container(
               color: BudgetTheme.isDarkMode(context)
-                  ? BudgetColors.primary600
-                  : BudgetColors.lightContainer,
+                  ? themeState.primaryColor[400]
+                  : themeState.primaryColor[100],
               padding: EdgeInsets.all(15),
               width: double.infinity,
               child: Row(
@@ -117,34 +108,20 @@ class DebtTile extends StatelessWidget {
                   bottomRight: Radius.circular(10.0),
                   topLeft: Radius.zero,
                   bottomLeft: Radius.circular(10.0)),
-              color: BudgetTheme.isDarkMode(context)
-                  ? BudgetColors.primary
-                  : BudgetColors.primary600,
+              color: themeState.primaryColor[600],
             ),
             child: Row(
               mainAxisAlignment: MainAxisAlignment.spaceBetween,
               children: [
                 Text('Completed: ',
-                    style: Theme.of(context)
-                        .textTheme
-                        .titleMedium!
-                        .copyWith(color: Colors.white)),
+                    style: Theme.of(context).textTheme.titleMedium!.copyWith(color: Colors.white)),
                 Text('0 %',
-                    style: Theme.of(context)
-                        .textTheme
-                        .titleMedium!
-                        .copyWith(color: Colors.white)),
+                    style: Theme.of(context).textTheme.titleMedium!.copyWith(color: Colors.white)),
               ],
             ),
           )
         ],
       ),
     );
-  }
-
-  Color _getColor(BuildContext context){
-    return BudgetTheme.isDarkMode(context)
-        ? BudgetColors.primary
-        : BudgetColors.lightContainer;
   }
 }
