@@ -1,11 +1,11 @@
 import 'package:budget_app/debt_payoff_planner/cubits/strategy_cubit/strategy_cubit.dart';
 import 'package:budget_app/debt_payoff_planner/models/debt_payoff_strategy.dart';
-import 'package:budget_app/utils/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:intl/intl.dart';
 
-import '../../../constants/colors.dart';
+import '../../../utils/theme/budget_theme.dart';
+import '../../../utils/theme/cubit/theme_cubit.dart';
 
 class PayoffSummary extends StatelessWidget {
   final DebtPayoffStrategy debtPayoffStrategy;
@@ -15,12 +15,13 @@ class PayoffSummary extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final textTheme = Theme.of(context).textTheme;
+    final themeState = context.read<ThemeCubit>().state;
     return BlocBuilder<StrategyCubit, StrategyState>(
       builder: (context, state) {
         return Card(
           color: BudgetTheme.isDarkMode(context)
-              ? BudgetColors.accentDark
-              : BudgetColors.accent,
+              ? themeState.primaryColor[400]
+              : themeState.primaryColor[100],
           margin: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
           child: Padding(
             padding: EdgeInsets.symmetric(vertical: 10, horizontal: 15),
@@ -33,30 +34,24 @@ class PayoffSummary extends StatelessWidget {
                         '${DateFormat('MM-dd-yyyy').format(debtPayoffStrategy.debtFreeDate)}',
                         style: textTheme.titleMedium!.copyWith(
                             fontWeight: FontWeight.bold,
-                            color: _getColor(context))),
-                    Text('DEBT FREE ON',
-                        style: textTheme.bodySmall!
-                            .copyWith(color: _getColor(context))),
+                            color: themeState.primaryColor[900])),
+                    Text('DEBT FREE ON', style: textTheme.bodySmall),
                   ],
                 ),
                 Column(
                   children: [
                     Text('${debtPayoffStrategy.totalDuration}',
-                        style: textTheme.titleMedium!.copyWith(
-                            fontWeight: FontWeight.bold,
-                            color: _getColor(context))),
-                    Text('DURATION', style: textTheme.bodySmall!
-            .copyWith(color: _getColor(context))),
+                        style: textTheme.titleMedium!
+                            .copyWith(fontWeight: FontWeight.bold, color: themeState.primaryColor[900])),
+                    Text('DURATION', style: textTheme.bodySmall),
                   ],
                 ),
                 Column(
                   children: [
                     Text('\$ ${debtPayoffStrategy.totalInterest}',
-                        style: textTheme.titleMedium!.copyWith(
-                            fontWeight: FontWeight.bold,
-                            color: _getColor(context))),
-                    Text('TOTAL INTEREST', style: textTheme.bodySmall!
-            .copyWith(color: _getColor(context))),
+                        style: textTheme.titleMedium!
+                            .copyWith(fontWeight: FontWeight.bold, color: themeState.primaryColor[900])),
+                    Text('TOTAL INTEREST', style: textTheme.bodySmall),
                   ],
                 ),
               ],
@@ -65,11 +60,5 @@ class PayoffSummary extends StatelessWidget {
         );
       },
     );
-  }
-
-  Color _getColor(BuildContext context) {
-    return BudgetTheme.isDarkMode(context)
-        ? BudgetColors.dark
-        : BudgetColors.primary;
   }
 }
