@@ -4,23 +4,24 @@ enum SubcategoriesStatus { loading, success, failure }
 
 class SubcategoriesState extends Equatable {
   final SubcategoriesStatus status;
+  final Budget? budget;
   final List<Subcategory> subcategories;
   final Category? category;
   final String? name;
-  final Subcategory? editSubcategory;
   final String? errorMessage;
 
   const SubcategoriesState({
     this.status = SubcategoriesStatus.loading,
+    this.budget,
     this.subcategories = const [],
     this.category,
     this.name,
-    this.editSubcategory,
     this.errorMessage,
   });
 
   SubcategoriesState copyWith({
     SubcategoriesStatus? status,
+    Budget? budget,
     List<Subcategory>? subcategories,
     Category? category,
     String? name,
@@ -29,10 +30,10 @@ class SubcategoriesState extends Equatable {
   }) {
     return SubcategoriesState(
         status: status ?? this.status,
+        budget: budget ?? this.budget,
         subcategories: subcategories ?? this.subcategories,
         category: category ?? this.category,
         name: name ?? this.name,
-        editSubcategory: editSubcategory ?? this.editSubcategory,
         errorMessage: errorMessage ?? this.errorMessage);
   }
 
@@ -41,11 +42,10 @@ class SubcategoriesState extends Equatable {
         status: this.status,
         subcategories: this.subcategories,
         category: this.category,
-        name: this.name,
-        editSubcategory: null);
+        name: this.name);
   }
 
   @override
   List<Object?> get props =>
-      [status, subcategories, category, name, editSubcategory, errorMessage];
+      [status, budget, subcategories, category, name, errorMessage];
 }
