@@ -1,4 +1,3 @@
-export 'widgets/widgets.dart';
 export 'bloc/transfer_bloc.dart';
 export 'view/view.dart';
-export 'models/models.dart';
+export 'widgets/widgets.dart';
