@@ -1,15 +1,16 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../../utils/theme/cubit/theme_cubit.dart';
 import '../bloc/transfer_bloc.dart';
 
 class AmountInput extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<TransferBloc, TransferState>(
-      builder: (context, state) {
-        return TextFormField(
-          initialValue: state.amount.value,
+    final themeState = context.watch<ThemeCubit>().state;
+    final amount = context.select((TransferBloc bloc) => bloc.state.amount);
+    return TextFormField(
+          initialValue: amount.value,
           onChanged: (amount) => context
               .read<TransferBloc>()
               .add(TransferAmountChanged(amount: amount)),
@@ -17,16 +18,14 @@ class AmountInput extends StatelessWidget {
           decoration: InputDecoration(
             icon: Icon(
               Icons.attach_money,
-              color: Colors.orangeAccent,
+              color: themeState.secondaryColor,
             ),
             border: OutlineInputBorder(),
             labelText: 'Amount',
             helperText: '',
             errorText:
-            state.amount.displayError != null ? 'invalid amount' : null,
+            amount.displayError != null ? 'invalid amount' : null,
           ),
         );
-      },
-    );
   }
 }
