@@ -1,30 +1,26 @@
-
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../../utils/theme/cubit/theme_cubit.dart';
 import '../bloc/transfer_bloc.dart';
 
 class NotesInputField extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<TransferBloc, TransferState>(
-      buildWhen: (previous, current) =>
-      previous.notes != current.notes,
-      builder: (context, state) {
-        return TextFormField(
-            initialValue: state.notes,
-            decoration: InputDecoration(
-              icon: Icon(
-                Icons.notes,
-                color: Colors.orangeAccent,
-              ),
-              border: OutlineInputBorder(),
-              labelText: 'Notes',
-            ),
-            onChanged: (notes) => context.read<TransferBloc>().add(
+    final themeState = context.watch<ThemeCubit>().state;
+    final notes = context.select((TransferBloc bloc) => bloc.state.notes);
+    return TextFormField(
+        initialValue: notes,
+        decoration: InputDecoration(
+          icon: Icon(
+            Icons.notes,
+            color: themeState.secondaryColor,
+          ),
+          border: OutlineInputBorder(),
+          labelText: 'Notes',
+        ),
+        onChanged: (notes) => context.read<TransferBloc>().add(
               TransferNotesChanged(notes: notes),
             ));
-      },
-    );
   }
 }
